Fixed another int/size_t problem that shows up on 64 bit systems (from Michael Ritzert" <michael@ritzert.de>).
git-svn-id: https://svn.code.sf.net/p/linux-gpib/code/trunk@1402 b61deed1-901b-0410-a6dd-e94ecbdaea89

+++ linux-gpib/driver/sys/ibread.c
@@ -36,7 +36,7 @@ int ibrd(gpib_board_t *board, uint8_t *buf, size_t length, int *end_flag, size_t
 {
 	ssize_t ret = 0;
 	int retval;
-	int bytes_read;
+	size_t bytes_read;
 
 	*nbytes = 0;
 	*end_flag = 0;
